renamed the module from sotb_wrapper to sotb.wrapper
