+++ setup.cfg
@@ -10,7 +10,7 @@ exclude = .github
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = sotb_wrapper/_version.py
-versionfile_build = sotb_wrapper/_version.py
+versionfile_source = sotb/wrapper/_version.py
+versionfile_build = sotb/wrapper/_version.py
 tag_prefix =
