+++ sotb/wrapper/_version.py
@@ -44,7 +44,7 @@ def get_config():
     cfg.style = "pep440"
     cfg.tag_prefix = ""
     cfg.parentdir_prefix = "None"
-    cfg.versionfile_source = "sotb_wrapper/_version.py"
+    cfg.versionfile_source = "sotb/wrapper/_version.py"
     cfg.verbose = False
     return cfg
