+++ sotb/wrapper/examples/plot_curves.py
